all in one c file
