--- MANIFEST.in
@@ -1,1 +1,0 @@
-include pytwed/twed.h
