--- pytwed/twed.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <math.h>
-
-double DTWEDL1d(int n_feats, double* arr1, int arr1_lgt, double* arr1_spec, double* arr2, int arr2_lgt, double* arr2_spec, double nu, double lambda, int degree);
